Refactor einiger doppelten Dateien
