+++ lama_app/lib/app/screens/task_screen.dart
@@ -7,10 +7,10 @@ import 'package:lama_app/app/event/task_events.dart';
 import 'package:lama_app/app/event/tts_event.dart';
 import 'package:lama_app/app/screens/task_type_screens/buchstabieren_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/clock_task_screen.dart';
-import 'package:lama_app/app/screens/task_type_screens/cloze_test_task_screen_stateful.dart';
+import 'package:lama_app/app/screens/task_type_screens/cloze_test_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/connect_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/equation_task_screen.dart';
-import 'package:lama_app/app/screens/task_type_screens/four_card_task_screen_stateful.dart';
+import 'package:lama_app/app/screens/task_type_screens/four_card_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/grid_select_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/mark_words_task_screen.dart';
 import 'package:lama_app/app/screens/task_type_screens/match_Category_task_screen.dart';
