+++ lama_app/lib/app/screens/task_type_screens/four_card_task_screen.dart
@@ -1,7 +1,10 @@
+import 'dart:developer';
+
 import 'package:bubble/bubble.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:http/http.dart';
 import 'package:lama_app/app/bloc/task_bloc.dart';
 import 'package:lama_app/app/event/task_events.dart';
 import 'package:lama_app/app/task-system/task.dart';
@@ -9,147 +12,206 @@ import 'package:lama_app/util/LamaColors.dart';
 import 'package:lama_app/util/LamaTextTheme.dart';
 import 'package:flutter_tts/flutter_tts.dart';
 
-/// [StatelessWidget] that contains the screen for the 4Cards TaskType.
-///
-/// Author: K.Binder
-class FourCardTaskScreen extends StatelessWidget {
-  final Task4Cards task;
-  final List<String?> answers = [];
-  final FlutterTts flutterTts = FlutterTts();
+import 'package:lama_app/app/bloc/taskBloc/tts_bloc.dart';
+import 'package:lama_app/app/event/tts_event.dart';
+import 'package:lama_app/app/state/tts_state.dart';
+import 'package:lama_app/app/state/QuestionText.dart';
 
-  //The constraints describe the space useable by the task.
-  //Its parent is always a container covering the whole available area
+class FourCardTaskScreenStateful extends StatefulWidget {
+  final Task4Cards task;
+  final List<String> answers = [];
   final BoxConstraints constraints;
 
-  FourCardTaskScreen(this.task, this.constraints) {
+  FourCardTaskScreenStateful(this.task, this.constraints) {
     answers.addAll(task.wrongAnswers);
-    answers.add(task.rightAnswer);
+    answers.add(task.rightAnswer!);
     print(answers.length);
     answers.shuffle();
+
+  }
+
+
+  @override
+  FourCards createState() {
+    FourCardTaskScreenStateful(task, constraints);
+    return FourCards(task, constraints, answers);
+  }
+
+
+}
+
+class FourCards extends State<FourCardTaskScreenStateful> {
+
+  final Task4Cards task;
+  final List<String> answers = [];
+  final BoxConstraints constraints;
+
+  String selectedAnswer = "";
+
+
+  FourCards(this.task, this.constraints, answers) {
+    this.answers.add(answers[0]);
+    this.answers.add(answers[1]);
+    this.answers.add(answers[2]);
+    this.answers.add(answers[3]);
   }
 
 
 
   @override
   Widget build(BuildContext context) {
-    // reads the question out loud
-    readquestion() async {
-      var text = task.question;
-      await flutterTts.setLanguage("de-De");
-      await flutterTts.setVolume(1.0);
-      await flutterTts.speak(text!);
-    }
-    readquestion();
-    return Column(children: [
-      Container(
-        height: (constraints.maxHeight / 100) * 40,
-        width: (constraints.maxWidth),
-        padding: EdgeInsets.all(25),
-        child: Container(
-            decoration: BoxDecoration(
-                borderRadius: BorderRadius.all(Radius.circular(50)),
-                gradient: LinearGradient(colors: [
-                  LamaColors.orangeAccent,
-                  LamaColors.orangePrimary
-                ]),
-                boxShadow: [
-                  BoxShadow(
-                      color: Colors.grey.withOpacity(0.5),
-                      spreadRadius: 5,
-                      blurRadius: 7,
-                      offset: Offset(0, 3))
-                ]),
-            child: Align(
-              child: Text(task.question!,
-                  textAlign: TextAlign.center,
-                  style: LamaTextTheme.getStyle(fontSize: 30)),
-            )),
-      ),
-      Container(
-        height: (constraints.maxHeight / 100) * 15,
-        padding: EdgeInsets.only(left: 15, right: 15),
-        child: Stack(children: [
-          Align(
-            alignment: Alignment.centerLeft,
-            child: Container(
-              padding: EdgeInsets.only(left: 75),
-              height: 50,
-              width: MediaQuery.of(context).size.width,
-              child: Bubble(
-                nip: BubbleNip.leftCenter,
-                child: Center(
-                  child: Text(
-                    task.lamaText!,
-                    style: TextStyle(fontSize: 15, fontWeight: FontWeight.bold),
+    String qlang;
+    task.questionLanguage == null || task.questionLanguage == "" ? qlang = "Deutsch" : qlang = "Englisch";
+
+    return BlocProvider(
+      create: (context) => TTSBloc(),
+      child: Column(children: [
+        Container(
+          height: (constraints.maxHeight / 100) * 40,
+          width: (constraints.maxWidth),
+          padding: EdgeInsets.all(25),
+          child: BlocBuilder<TTSBloc, TTSState>(
+            builder: (context, TTSState state) {
+              if (state is EmptyTTSState) {
+                //log('task.questionLanguage: ${task.questionLanguage}');
+                //log('task.answerLaguage: ${task.answerLaguage}');
+                context.read<TTSBloc>().add(AnswerOnInitEvent(task.question!,qlang));
+                QuestionText.setText(task.question!, qlang);
+              }
+              return Container(
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(Radius.circular(50)),
+                    gradient: LinearGradient(colors: [
+                      LamaColors.orangeAccent,
+                      LamaColors.orangePrimary
+                    ]),
+                    boxShadow: [
+                      BoxShadow(
+                          color: Colors.grey.withOpacity(0.5),
+                          spreadRadius: 5,
+                          blurRadius: 7,
+                          offset: Offset(0, 3))
+                    ]),
+                child: Align(
+                  child: InkWell(
+                    child: Text(task.question!,
+                        textAlign: TextAlign.center,
+                        style: LamaTextTheme.getStyle(fontSize: 30)
+                    ),
+                  ),
+                ),
+              );
+            },
+          ),
+        ),
+        Container(
+          height: (constraints.maxHeight / 100) * 15,
+          padding: EdgeInsets.only(left: 15, right: 15),
+          child: Stack(children: [
+            Align(
+              alignment: Alignment.centerLeft,
+              child: Container(
+                padding: EdgeInsets.only(left: 75),
+                height: 50,
+                width: MediaQuery
+                    .of(context)
+                    .size
+                    .width,
+                child: Bubble(
+                  nip: BubbleNip.leftCenter,
+                  child: Center(
+                    child: Text(
+                      "Tippe einmal, um die Antwort oder Frage anzuhören.",
+                      style: TextStyle(fontSize: 15,
+                          fontWeight: FontWeight.bold),
+                    ),
                   ),
                 ),
               ),
             ),
-          ),
-          Align(
-            alignment: Alignment.centerLeft,
-            child: SvgPicture.asset(
-              "assets/images/svg/lama_head.svg",
-              semanticsLabel: "Lama Anna",
-              width: 75,
-            ),
-          ),
-        ]),
-      ),
-      Container(
-          height: (constraints.maxHeight / 100) * 45,
-          child: Padding(
-              padding: const EdgeInsets.only(
-                top: 10,
-                left: 5,
-                right: 5,
+            Align(
+              alignment: Alignment.centerLeft,
+              child: SvgPicture.asset(
+                "assets/images/svg/lama_head.svg",
+                semanticsLabel: "Lama Anna",
+                width: 75,
               ),
-              child: GridView.builder(
-                  gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                    crossAxisCount: 2,
-                    childAspectRatio: 1.6 / 1,
-                    mainAxisSpacing: 10,
-                    crossAxisSpacing: 10,
-                  ),
-                  itemCount: 4,
-                  itemBuilder: (context, index) =>
-                      _buildCards(context, index)))),
+            ),
+          ]),
+        ),
+        Container(
+            height: (constraints.maxHeight / 100) * 45,
+            child: Padding(
+                padding: const EdgeInsets.only(
+                  top: 10,
+                  left: 5,
+                  right: 5,
+                ),
+                child: GridView.builder(
+                    gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                      crossAxisCount: 2,
+                      childAspectRatio: 1.6 / 1,
+                      mainAxisSpacing: 10,
+                      crossAxisSpacing: 10,
+                    ),
+                    itemCount: 4,
+                    itemBuilder: (context, index) =>
+                        _buildCards(context, index)))),
 
-    ]);
-
+      ]),
+    );
   }
 
-  ///Returns one of the four cards that contain the different answers as [Widget].
   Widget _buildCards(context, index) {
+    String alang;
+    task.answerLanguage == null || task.answerLanguage == "" ? alang = "Deutsch" : alang = task.answerLanguage!;
+    //debugPrint(index);
     Color color =
-        index % 3 == 0 ? LamaColors.greenAccent : LamaColors.blueAccent;
-    return Container(
-      height: 50,
-      decoration: BoxDecoration(
-          borderRadius: BorderRadius.all(Radius.circular(20)),
-          color: color,
-          boxShadow: [
-            BoxShadow(
+    index % 3 == 0 ? LamaColors.greenAccent : LamaColors.blueAccent;
 
-                color: Colors.grey.withOpacity(0.5),
-                spreadRadius: 1,
-                blurRadius: 7,
-                offset: Offset(0, 3))
-          ]),
-      child: InkWell(
-        onTap: () => BlocProvider.of<TaskBloc>(context)
-            .add(AnswerTaskEvent(answers[index])),
-        child: Padding(
-          padding: EdgeInsets.all(5),
-          child: Center(
-            child: Text(
-              answers[index]!,
-              textAlign: TextAlign.center,
-              style: LamaTextTheme.getStyle(),
+    return BlocBuilder<TTSBloc, TTSState>(
+      builder: (context, state) {
+        return Container(
+          height: 50,
+          decoration: BoxDecoration(
+              borderRadius: BorderRadius.all(Radius.circular(20)),
+              color: state is VoiceAnswerTtsState &&
+                  state.selectedAnswer == answers[index] ? LamaColors
+                  .purpleAccent : color,
+              boxShadow: [
+                BoxShadow(
+                    color: Colors.grey.withOpacity(0.5),
+                    spreadRadius: 1,
+                    blurRadius: 7,
+                    offset: Offset(0, 3))
+              ]),
+          child: InkWell(
+            onTap: () {
+              if ( selectedAnswer != answers[index]) {
+                //log('data: $selectedAnswer');
+                BlocProvider.of<TTSBloc>(context)
+                .add(ClickOnAnswer(answers[index], index, alang));
+                selectedAnswer = answers[index];
+              } else {
+                BlocProvider.of<TaskBloc>(context)
+                    .add(AnswerTaskEvent(answers[index]));
+                BlocProvider.of<TTSBloc>(context).
+                add(SetDefaultEvent());
+              }
+            },
+            child: Padding(
+              padding: EdgeInsets.all(5),
+              child: Center(
+                child: Text(
+                  answers[index],
+                  textAlign: TextAlign.center,
+                  style: LamaTextTheme.getStyle(),
+                ),
+              ),
             ),
           ),
-        ),
-      ),
+        );
+      },
     );
   }
 }
